+++ pingche_project/src/main/java/com/yf/pingche/entity/Msg.java
@@ -30,9 +30,9 @@ public class Msg implements Serializable {
 private static final long serialVersionUID=1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
-    private Integer uid;
+    private Long uid;
 
     private String content;
 
@@ -44,7 +44,7 @@ private static final long serialVersionUID=1L;
 
     private String url;
 
-    private Integer fid;
+    private Long fid;
 
 
 }
